Some parameter mismatch fixation
I just matched the input parameters for account management-related codes. The user profile Controller and Entity have not been done yet. Can match to this code style after checking the account CRUDS are working correctly.

I think most of the function can be a static function. Since it is just simply used for the "Connection", it does not need to create an object that seriously.